MEMAP: Checking for alignment issues when doing a 16bit or 32bit read or write.
The ARM debug interface v5 states that accessing an address that is not
aligned to the access size is "implementation defined".

I think it's better to not allow this, as otherwise we would have to
program in support for each implementation method, and work out which
implementations support which method.