DEBUG: Don't re-read the CPUID register for every target.
Previously we iterated through the list of CPUIDs, read the
specified address and then compared it to the default value.

However all the addresses were the same. It seems silly to read the same
register N times.

I've changed this to use a define of the address, reading once, and
comparing to each entry in the CPUID list.

This is much faster especially if the SWD baudrate is slow or there are
lots of supported CPUIDs.

If in the future we find a chip that has it's CPUID in a different
location then we will need to change this code again. Maybe to have a
list of addresses, and for each address a list of pairs of name and
default value.