MEMAP: Fixing bug with auto TAR increment (Issue #15).
The TAR register can be auto incremented after every write or read from
DRW. However this is only guaranteed to work for the lowest 10 bits of
TAR. Above that it's implementation defined.

On some targets it's just the 10 bits, meaning that something unknown
<implementation defined> happens if you go over a !KB boundary.

The previous code dealt with this, by rewriting the TAR register every
1KB. However that doesn't work if you don't start your writes 1KB
aligned.

For example, if you write 1KB starting at address 0x200, the first 512
bytes write OK, after that it's implementation defined. On my target it
wrapped to 0 and wrote the last 512 bytes there.

My fix re-writes the TAR register whenever you hit a 1KB boundary,
regardless of start address.